requirements.txt
textblog
nltk.stem
collections
sys
pprint
re
